scrapers/requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
scrapy==2.11.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Redis
redis==5.0.1
celery==5.3.4

# Data processing
pandas==2.1.4
numpy==1.26.2

# Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Logging and monitoring
python-json-logger==2.0.7
sentry-sdk==1.39.1

# HTTP and networking
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0
pydantic==2.5.2
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1